# improve performance of setindex! on IdDict