@KristofferC 2019-08-21
@nanosoldier `runbenchmarks("collection", vs = ":master")`