@nanosoldier 2019-08-21
[Your benchmark job](https://github.com/JuliaLang/julia/pull/33009#issuecomment-523473579) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/d104df8dae58871d1792273f68ab4396dbaffea5/3d6240f_vs_6723492/report.md). cc @ararslan