@KristofferC 2019-08-22
Performance looks good but would be good to know if the now dynamic call to convert is expected, cc @vtjnash, @JeffBezanson .